Move out result to storages/result
