+++ Makefile
@@ -16,12 +16,11 @@ setup:
 fmt:
 	cargo fmt --all
 
-machete:
-	cargo -Z sparse-registry machete
-
 lint:
 	cargo fmt --all
 	cargo clippy --workspace --all-targets -- -D warnings
+	# Check unused deps(make setup to install)
+	cargo -Z sparse-registry machete
 
 	# Cargo.toml file formatter(make setup to install)
 	taplo fmt
