+++ src/query/service/src/servers/http/v1/download.rs
@@ -22,10 +22,10 @@ use common_meta_types::StageFileFormatType;
 use common_planner::extras::Extras;
 use common_planner::ReadDataSourcePlan;
 use common_planner::SourceInfo;
+use common_storages_result::ResultTable;
 use futures::StreamExt;
 
 use crate::sessions::QueryContext;
-use crate::storages::result::ResultTable;
 use crate::storages::Table;
 use crate::stream::DataBlockStream;
