+++ src/query/service/src/servers/http/v1/query/http_query.rs
@@ -23,6 +23,7 @@ use common_base::base::tokio::sync::RwLock;
 use common_base::base::TrySpawn;
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_storages_result::BlockBuffer;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -44,7 +45,6 @@ use crate::servers::http::v1::HttpQueryManager;
 use crate::sessions::QueryAffect;
 use crate::sessions::SessionType;
 use crate::sessions::TableContext;
-use crate::storages::result::block_buffer::BlockBuffer;
 
 #[derive(Deserialize, Debug)]
 pub struct HttpQueryRequest {
