+++ src/query/service/tests/it/storages/result/result_table.rs
@@ -26,10 +26,10 @@ use common_exception::Result;
 use common_meta_types::UserIdentity;
 use common_planner::ReadDataSourcePlan;
 use common_planner::SourceInfo;
+use common_storages_result::ResultQueryInfo;
+use common_storages_result::ResultTable;
+use common_storages_result::ResultTableWriter;
 use databend_query::sessions::TableContext;
-use databend_query::storages::result::ResultQueryInfo;
-use databend_query::storages::result::ResultTable;
-use databend_query::storages::result::ResultTableWriter;
 use databend_query::storages::Table;
 use databend_query::stream::DataBlockStream;
 use futures::TryStreamExt;
