+++ src/query/storages/factory/Cargo.toml
@@ -12,30 +12,13 @@ doctest = false
 test = false
 
 [dependencies]
-common-base = { path = "../../../common/base" }
 common-catalog = { path = "../../catalog" }
 common-config = { path = "../../config" }
-common-datablocks = { path = "../../datablocks" }
-common-datavalues = { path = "../../datavalues" }
 common-exception = { path = "../../../common/exception" }
-common-fuse-meta = { path = "../fuse-meta" }
 common-meta-app = { path = "../../../meta/app" }
-common-meta-types = { path = "../../../meta/types" }
-common-pipeline-core = { path = "../../pipeline/core" }
-common-pipeline-transforms = { path = "../../pipeline/transforms" }
-common-planner = { path = "../../planner" }
 common-storages-fuse = { path = "../fuse" }
 common-storages-index = { path = "../index" }
 common-storages-preludes = { path = "../preludes" }
 common-storages-view = { path = "../view" }
-common-streams = { path = "../../streams" }
 
-async-trait = { version = "0.1.57", package = "async-trait-fn" }
-backon = "0.2"
-futures = "0.3.24"
-opendal = { version = "0.19", features = ["layers-retry", "layers-tracing", "layers-metrics", "compress"] }
 parking_lot = "0.12.1"
-serde = { workspace = true }
-serde_json = { workspace = true }
-tracing = "0.1.36"
-uuid = { version = "1.1.2", features = ["serde", "v4"] }
