+++ src/query/storages/factory/src/lib.rs
@@ -14,7 +14,6 @@
 
 pub use common_storages_fuse as fuse;
 pub use common_storages_index as index;
-pub mod result;
 mod storage_factory;
 
 pub use common_catalog::table::NavigationPoint;
