+++ src/query/storages/result/src/lib.rs
@@ -12,13 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod block_buffer;
+mod block_buffer;
 mod result_locations;
 mod result_table;
 mod result_table_sink;
 mod result_table_source;
 mod writer;
 
+pub use block_buffer::BlockBuffer;
+pub use block_buffer::BlockBufferWriterMemOnly;
+pub use block_buffer::BlockBufferWriterWithResultTable;
 pub use result_table::ResultQueryInfo;
 pub use result_table::ResultTable;
 pub use result_table_sink::ResultTableSink;
